% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_to_sf.R
\name{entropy_to_sf}
\alias{entropy_to_sf}
\title{Convert local entropy results to an sf object}
\usage{
entropy_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a \code{segreg} object returned by
\code{\link{measure_segregation}}.}

\item{bandwidths}{numeric vector. Optional subset of bandwidths to include.
When empty (default), all bandwidths are returned.}
}
\value{
a spatial sf object with columns \code{id}, \code{bw}, and
    \code{entropy}.
}
\description{
Convert local entropy results to an sf object
}
\examples{

library(sf)
library(segregr)

marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg",
                                  package = "segregr"))
segregation <- measure_segregation(marilia_sf)
entropy <- entropy_to_sf(segregation)

if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot(data = entropy) +
    ggplot2::geom_sf(ggplot2::aes(fill = entropy)) +
    ggplot2::scale_fill_distiller(palette = "Spectral")
}
}
