% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_isolation_matrix.R
\name{exposure_isolation_matrix}
\alias{exposure_isolation_matrix}
\title{Convert global exposure and isolation to tabular form}
\usage{
exposure_isolation_matrix(segregation_results)
}
\arguments{
\item{segregation_results}{a \code{segreg} object returned by
\code{\link{measure_segregation}}.}
}
\value{
a data frame with one row per group and one column per group,
    where off-diagonal cells are pairwise Exposure indices and diagonal
    cells are Isolation indices.
}
\description{
Convert global exposure and isolation to tabular form
}
\examples{

library(sf)
library(segregr)

marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg",
                                  package = "segregr"))
segregation <- measure_segregation(marilia_sf)
exposure_isolation_matrix(segregation)
}
