% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_to_sf.R
\name{exposure_to_sf}
\alias{exposure_to_sf}
\title{Convert local exposure results to an sf object}
\usage{
exposure_to_sf(segregation_results)
}
\arguments{
\item{segregation_results}{a \code{segreg} object returned by
\code{\link{measure_segregation}}.}
}
\value{
a spatial sf object with columns \code{id}, \code{bw},
    \code{group_a}, \code{group_b}, and \code{exposure}.
}
\description{
Convert local exposure results to an sf object
}
\examples{

library(sf)
library(segregr)

marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg",
                                  package = "segregr"))
segregation <- measure_segregation(marilia_sf)
exposure <- exposure_to_sf(segregation)

if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot(data = exposure) +
    ggplot2::geom_sf(ggplot2::aes(fill = exposure), color = NA) +
    ggplot2::scale_fill_distiller(palette = "Spectral") +
    ggplot2::facet_grid(group_a ~ group_b) +
    ggplot2::theme_void()
}
}
