% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_to_sf.R
\name{h_to_sf}
\alias{h_to_sf}
\title{Title}
\usage{
h_to_sf(segregation_results)
}
\arguments{
\item{segregation_results}{a SEGREG object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local results of Theil's index H
}
\description{
Title
}
