% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolation_exposure_matrix.R
\name{isolation_exposure_matrix}
\alias{isolation_exposure_matrix}
\title{Title}
\usage{
isolation_exposure_matrix(segregation_results)
}
\arguments{
\item{segregation_results}{a SEGREG object containing the results of a
call to measure_segregation().}
}
\value{
a data.frame with global exposure and isolation results in tabular form
}
\description{
Title
}
