% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_to_sf.R
\name{dissimilarity_to_sf}
\alias{dissimilarity_to_sf}
\title{Convert local dissimilarity results to SF}
\usage{
dissimilarity_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}

\item{bandwidths}{Attributes to be used. Default is all.}
}
\value{
a spatial sf object with local dissimilarity results
}
\description{
Convert local dissimilarity results to SF
}
\examples{
require(tmap)
# load sample data from package segregr
marilia_sf <- sf::st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- segregr::measure_segregation(marilia_sf)

# export local dissimilarity results
dissimilarity <- segregr::dissimilarity_to_sf(segregation)

# plot local dissimilarity
tm_shape(dissimilarity) +
  tm_polygons("dissimilarity", n = 3) +
  tm_layout(legend.position=c(0.1,0.1), scale = 0.5)
}
