% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_to_sf.R
\name{dissimilarity_to_sf}
\alias{dissimilarity_to_sf}
\title{Title}
\usage{
dissimilarity_to_sf(segregation_results)
}
\arguments{
\item{segregation_results}{a SEGREG object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local dissimilarity results
}
\description{
Title
}
