% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_to_sf.R
\name{entropy_to_sf}
\alias{entropy_to_sf}
\title{Convert local entropy results to SF}
\usage{
entropy_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}

\item{bandwidths}{Bands to be used.}
}
\value{
A spatial sf object with local entropy results.
}
\description{
Convert local entropy results to SF
}
\examples{
require(tmap)
# load sample data from package segregr
marilia_sf <- sf::st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- segregr::measure_segregation(marilia_sf)

# export local entropy results
entropy <- segregr::entropy_to_sf(segregation)

# plot local entropy
tm_shape(entropy) +
  tm_polygons("entropy", n = 3) +
  tm_layout(legend.position=c(0.1, 0.1), scale = 0.5)
}
