% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_to_sf.R
\name{entropy_to_sf}
\alias{entropy_to_sf}
\title{Convert local entropy results to SF}
\usage{
entropy_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local entropy results
}
\description{
Convert local entropy results to SF
}
\examples{

library("sf")
library("ggplot2")
library("segregr")

# load sample data from package segregr
marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- measure_segregation(marilia_sf)

# export local entropy results
entropy <- entropy_to_sf(segregation)

# plot local entropy
ggplot(data = entropy) +
  geom_sf(aes(fill = entropy)) +
  scale_fill_distiller(palette = "Spectral")
}
