% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_to_sf.R
\name{exposure_to_sf}
\alias{exposure_to_sf}
\title{Convert local exposure results to SF}
\usage{
exposure_to_sf(segregation_results)
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local exposure results
}
\description{
Convert local exposure results to SF
}
\examples{
# load sample data from package segregr
marilia_sf <- sf::st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- segregr::measure_segregation(marilia_sf)

# export local isolation results
exposure <- segregr::exposure_to_sf(segregation)

# plot local exposure
#ggplot(data = exposure) +
#   geom_sf(aes(fill = exposure), color = NA) +
#  scale_fill_distiller(palette = "Spectral") +
#   facet_grid(group_a ~ group_b) +
#  theme_void()
}
