% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_to_sf.R
\name{exposure_to_sf}
\alias{exposure_to_sf}
\title{Title}
\usage{
exposure_to_sf(segregation_results)
}
\arguments{
\item{segregation_results}{a SEGREG object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local exposure results
}
\description{
Title
}
