% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_metrics_to_df.R
\name{global_metrics_to_df}
\alias{global_metrics_to_df}
\title{Title}
\usage{
global_metrics_to_df(segregation_results, bandwidths = c())
}
\arguments{
\item{bandwidths}{}
}
\value{

}
\description{
Title
}
