% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_to_sf.R
\name{h_to_sf}
\alias{h_to_sf}
\title{Convert local results of information theory index H to SF}
\usage{
h_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}

\item{bandwidths}{Attributes to be used. Default is all.}
}
\value{
a spatial sf object with local index H results
}
\description{
Convert local results of information theory index H to SF
}
\examples{
require(tmap)
# load sample data from package segregr
marilia_sf <- sf::st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- segregr::measure_segregation(marilia_sf)

# export local index h results
index_h <- segregr::h_to_sf(segregation)

tm_shape(index_h) +
  tm_polygons("h", n = 3) +
  tm_layout(legend.position=c(0.1, 0.1), scale = 0.5)
}
