% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_to_sf.R
\name{h_to_sf}
\alias{h_to_sf}
\title{Convert local results of information theory index H to SF}
\usage{
h_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local index H results
}
\description{
Convert local results of information theory index H to SF
}
\examples{

library("sf")
library("ggplot2")
library("segregr")

# load sample data from package segregr
marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- measure_segregation(marilia_sf)

# export local index h results
index_h <- h_to_sf(segregation)

# plot local index h
ggplot(data = index_h) +
  geom_sf(aes(fill = h)) +
  scale_fill_distiller(palette = "Spectral")
}
