% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolation_to_sf.R
\name{isolation_to_sf}
\alias{isolation_to_sf}
\title{Convert local isolation results to SF}
\usage{
isolation_to_sf(segregation_results)
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}
}
\value{
a spatial sf object with local isolation results
}
\description{
Convert local isolation results to SF
}
\examples{
require(tmap)

# load sample data from package segregr
marilia_sf <- sf::st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- segregr::measure_segregation(marilia_sf)

# export local isolation results
isolation <- segregr::isolation_to_sf(segregation)

# plot local isolation
tm_shape(isolation) +
  tm_polygons("isolation", n = 3) +
  tm_layout(legend.position=c(0.1,0.1), scale = 0.5)
}
