% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_metrics_to_sf.R
\name{local_metrics_to_sf}
\alias{local_metrics_to_sf}
\title{Title}
\usage{
local_metrics_to_sf(segregation_results, bandwidths = c())
}
\arguments{
\item{segregation_results}{a segreg object containing the results of a
call to measure_segregation().}

\item{bandwidths}{}
}
\value{

}
\description{
Title
}
