% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_segregation.R
\name{measure_segregation}
\alias{measure_segregation}
\title{Calculate segregation metrics}
\usage{
measure_segregation(
  data,
  id_field = "id",
  distance_method = "geodist",
  weight_function = "gaussian",
  bandwidths = 0
)
}
\arguments{
\item{data}{a spatial sf object of the study area divided into areal
units (such as census tracts), containing the columns 'id', 'geometry',
and one column per population group indicating that group's population in
the local areal unit.}

\item{id_field}{character. The name of the identificator column in the input
data. Defaults to 'id'.}

\item{distance_method}{The method to compute distance between spatial objects.
Default is "geodist".}

\item{weight_function}{The weight function. Default is "gaussian".}

\item{bandwidths}{numeric. A bandwidth, in meters, used to set the scale of
analysis for spatial segregation measurement. When set to 0, aspatial
metrics will be calculated.}
}
\value{
a segreg object, a list containing the input spatial data and the
    results of the segregation metrics.
}
\description{
Calculate segregation metrics
}
\examples{
# load sample data from package segregr
marilia_sf <- sf::st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

# calculate segregation metrics
segregation <- segregr::measure_segregation(marilia_sf)

# global dissimilarity index
segregation$D

# global entropy
segregation$E

# global information theory index H
segregation$H
}
